修改目录基本结构
