+++ .idea/artifacts/BookManagerSystem_war.xml
@@ -1,0 +1,8 @@
+<component name="ArtifactManager">
+  <artifact type="war" name="BookManagerSystem:war">
+    <output-path>$PROJECT_DIR$/target</output-path>
+    <root id="archive" name="BookManagerSystem.war">
+      <element id="artifact" artifact-name="BookManagerSystem:war exploded" />
+    </root>
+  </artifact>
+</component>
