+++ .idea/compiler.xml
@@ -9,5 +9,8 @@
         <module name="BookManagerSystem" />
       </profile>
     </annotationProcessing>
+    <bytecodeTargetLevel>
+      <module name="BookManagerSystem" target="1.5" />
+    </bytecodeTargetLevel>
   </component>
 </project>
