+++ .idea/libraries/Maven__net_sf_ezmorph_ezmorph_1_0_6.xml
@@ -1,0 +1,13 @@
+<component name="libraryTable">
+  <library name="Maven: net.sf.ezmorph:ezmorph:1.0.6">
+    <CLASSES>
+      <root url="jar://$MAVEN_REPOSITORY$/net/sf/ezmorph/ezmorph/1.0.6/ezmorph-1.0.6.jar!/" />
+    </CLASSES>
+    <JAVADOC>
+      <root url="jar://$MAVEN_REPOSITORY$/net/sf/ezmorph/ezmorph/1.0.6/ezmorph-1.0.6-javadoc.jar!/" />
+    </JAVADOC>
+    <SOURCES>
+      <root url="jar://$MAVEN_REPOSITORY$/net/sf/ezmorph/ezmorph/1.0.6/ezmorph-1.0.6-sources.jar!/" />
+    </SOURCES>
+  </library>
+</component>
