+++ .idea/libraries/Maven__net_sf_json_lib_json_lib_jdk15_2_4.xml
@@ -1,0 +1,13 @@
+<component name="libraryTable">
+  <library name="Maven: net.sf.json-lib:json-lib:jdk15:2.4">
+    <CLASSES>
+      <root url="jar://$MAVEN_REPOSITORY$/net/sf/json-lib/json-lib/2.4/json-lib-2.4-jdk15.jar!/" />
+    </CLASSES>
+    <JAVADOC>
+      <root url="jar://$MAVEN_REPOSITORY$/net/sf/json-lib/json-lib/2.4/json-lib-2.4-javadoc.jar!/" />
+    </JAVADOC>
+    <SOURCES>
+      <root url="jar://$MAVEN_REPOSITORY$/net/sf/json-lib/json-lib/2.4/json-lib-2.4-sources.jar!/" />
+    </SOURCES>
+  </library>
+</component>
