--- src/main/bookManager/java/system/controller/BookController.java
@@ -1,11 +1,0 @@
-package bookManager.java.system.controller;
-
-import bookManager.java.common.support.BaseController;
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.RequestMapping;
-
-@Controller
-@RequestMapping("book")
-public class BookController extends BaseController {
-
-}
