--- src/main/bookManager/java/system/service/BookService.java
@@ -1,17 +1,0 @@
-package bookManager.java.system.service;
-
-import bookManager.java.system.domain.Book;
-
-import java.util.List;
-
-public interface BookService {
-    public List<Book> getBookList();
-
-    public Book getBookById(String bookId);
-
-    public int insertBook(Book book);
-
-    public int dropBookById(String bookId);
-
-    public int updateBook(Book book);
-}
