+++ src/main/java/system/controller/BookController.java
@@ -1,0 +1,11 @@
+package system.controller;
+
+import common.support.BaseController;
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.RequestMapping;
+
+@Controller
+@RequestMapping("book")
+public class BookController extends BaseController {
+
+}
