+++ src/main/resource/log4j.properties
